! Varios cambios
- log conversion pattern actualizado (fecha)
- cambio en la ubicacion de los logs en la configuracion de la app
+ nuevo shell script para confgurar la app como servicio en linux
! muevo los shell scripts a una carpeta bin

+++ src/main/java/com/proeza/mqttlogger/parser/Payload2SimpleLogParser.java
@@ -31,11 +31,11 @@ public class Payload2SimpleLogParser implements IMqttPayloadParser<String> {
 		return builder.toString().trim();
 	}
 
-	private int getChunkSize (String chunk) {
+	private short getChunkSize (String chunk) {
 		if (SIZE_WILDCARD.equals(chunk)) {
-			return Integer.SIZE;
+			return Short.SIZE;
 		} else {
-			return Integer.parseInt(chunk);
+			return Short.parseShort(chunk);
 		}
 	}
 }
